+++ crates/kira/src/sound/streaming/sound/test.rs
@@ -497,6 +497,7 @@ fn stops_if_waiting_on_missing_clock() {
 	);
 	sound.on_start_processing();
 	assert_eq!(handle.state(), PlaybackState::Stopped);
+	assert!(sound.finished());
 }
 
 /// Tests that a `StreamingSound` that had its start time set to a clock time and already